Add files via upload
This is a search application that takes as input an unsorted integer array and returns the original index value of the searched number, or -1 if the number is not in the original array. This is a workaround for the problem of binary search methods being unable to work with unsorted arrays. The numbers in the original array are appended with a decimal value that preserves the original index value. The unsorted array is then sorted and searched using the binary method. A minimum amount of post-processing allows the original index location to be extracted and returned as the output of the binary search. 

To the best of my knowledge, this should be O(n log(n)) time complexity, though it has not been tested. It should be more efficient overall than a linear search, which is the normal method for searching an unsorted array and has a complexity of O(n^2)

This app or any portion thereof is free to use for academic purposes. I do ask that you contact me for permission if using all or any portion for commercial applications. Thanks!